+++ src/Quill.Cli/CliJsonContext.cs
@@ -12,6 +12,7 @@ namespace Quill.Cli;
 [JsonSerializable(typeof(List<SearchResult>))]
 [JsonSerializable(typeof(List<CommentResult>))]
 [JsonSerializable(typeof(List<PullRequestResult>))]
+[JsonSerializable(typeof(List<PullRequestRevisionResult>))]
 [JsonSerializable(typeof(List<PullRequestThreadResult>))]
 [JsonSerializable(typeof(PullRequestViewResult))]
 [JsonSerializable(typeof(TreeNode))]